wrote a lot of analysis, with some math too!

+++ MonitorDriver.cpp
@@ -67,13 +67,13 @@ void* Schedule(void* Mon){
             clock_t start = clock();
             int N; 
             double T;
-            long turns, distance;
+            int  turns, distance;
             M->Request(track, duration, N, T, turns, distance);
             //M->NumberOfRequests(N);
             TimingFP = fopen("TimePerRequest.txt", "a");
             //fprintf(TimingFP, "%d %16.14f\n", N, (clock()/(double)CLOCKS_PER_SEC) -
             //                            start/(double)CLOCKS_PER_SEC);
-            fprintf(TimingFP, "%d %16.14f %ld %ld\n", N, T, turns, distance);
+            fprintf(TimingFP, "%d %16.14f %d %d\n", N, T, turns, distance);
 
             fclose(TimingFP);
             usleep(rand() % WAIT_TIME);
